fix math on handled rim
Total number of stitches on rim needed to be 1 less than total for that row in order to leave room for joining with sl st.

Therefore the number of stitches on the first spacer (which will always be odd) needed to be 1 SHORTER than the other (achieved by flooring both instances of `spacer/2` instead of flooring one and ceiling-ing the other).